소수점 처리
소수점 이하 2째자리까지만 표시하도록 변경
(3째자리에서 반올림)

+++ OTT/Group1ViewController.swift
@@ -13,7 +13,7 @@ class Group1ViewController: UIViewController , UITableViewDataSource,UITableView
     var valueToPass:Int = 0
     var valueToTitle: String = String()
     var marrGroupData : NSMutableArray!
-    var averageValue:Double = 0
+    var averageValue:Float = 0
     @IBOutlet weak var tbGroupData: UITableView!
     
     override func viewDidLoad() {
